refactor: update import paths for TimelineExperienceItem and experienceList; clean up unused imports in TabProjects
